refactor: move account schema to core
